# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MOTHERSHIP Multi-Agent Governance API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /execution/config:
    get:
      operationId: getExecutionConfig
      tags: [execution]
      summary: Get the (redacted) global execution config — wallet address, paper-mode flag, risk caps. Never returns the private key.
      responses:
        "200":
          description: Current execution config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionConfig"
    put:
      operationId: updateExecutionConfig
      tags: [execution]
      summary: Update execution config. Send `privateKey` as plaintext only over the local devloop — it is encrypted at rest with AES-256-GCM (key derived from SESSION_SECRET via HKDF) and never echoed back. Send empty string to clear.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExecutionConfigBody"
      responses:
        "200":
          description: Updated config (redacted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionConfig"
  /missions/{id}/execute:
    post:
      operationId: executeMission
      tags: [execution]
      summary: Submit the Queen's FINAL_VERDICT trade plan to Hyperliquid (or paper-simulate it). Places market entry + reduce-only stop + 50/50 reduce-only TP1/TP2 in a single grouped order.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Execution submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionPlacementResult"
        "400":
          description: Validation / risk-cap failure
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
        "409":
          description: Mission is already executing or completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  error: { type: string }
  /missions/{id}/execute/kill:
    post:
      operationId: killMissionExecution
      tags: [execution]
      summary: KILL switch — cancel all resting orders for the mission and flat-close any open position (real or paper).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Kill executed (may be a no-op if nothing is open)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionKillResult"
  /ledger:
    get:
      operationId: listLedgerEntries
      tags: [ledger]
      summary: List ledger entries (filterable by mission, action, time range). Read-only — the ledger is append-only.
      parameters:
        - name: missionId
          in: query
          required: false
          schema: { type: integer }
        - name: action
          in: query
          required: false
          schema:
            type: string
            enum: [submit, fill, cancel, kill, close, error]
        - name: after
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: before
          in: query
          required: false
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 5000 }
        - name: offset
          in: query
          required: false
          schema: { type: integer, minimum: 0 }
      responses:
        "200":
          description: Filtered ledger page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerPage"
...
  /ledger/missions/{id}:
    get:
      operationId: getMissionLedger
      tags: [ledger]
      summary: List ledger entries for a single mission.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Mission ledger page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerPage"
  /ledger/missions/{id}/verify:
    get:
      operationId: verifyMissionLedger
      tags: [ledger]
      summary: Verify the full chain and report how many entries belong to this mission.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Verify result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LedgerVerifyResult"
...
  /missions/{id}/execution:
    get:
      operationId: getMissionExecution
      tags: [execution]
      summary: Get the current execution state for a mission — every order leg plus the open (or last) position with refreshed unrealized PnL.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Execution snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionExecutionSnapshot"
components:
  schemas:
    LedgerEntry:
      type: object
      description: One link in the SHA-256 hash chain. `hash = SHA-256(prevHash + canonicalJSON(body))` where body is every field except `prevHash` and `hash`.
      properties:
        index: { type: integer, description: 0-based position in the chain }
        ts: { type: string, format: date-time }
        missionId: { type: integer, nullable: true }
        action:
          type: string
          enum: [submit, fill, cancel, kill, close, error]
        source: { type: string, description: e.g. "executor", "executor.paper", "kill" }
        missionObjective: { type: string, nullable: true }
        verdictRef:
          type: object
          nullable: true
          properties:
            stance: { type: string }
            symbol: { type: string, nullable: true }
            bias: { type: string, nullable: true }
            cycle: { type: integer, nullable: true }
            packetId:
              type: integer
              nullable: true
              description: Reasoning-packet row ID of the Queen FINAL_VERDICT that authorised this execution.
          required: [stance]
        tradePlan:
          nullable: true
          description: Snapshot of the Queen FINAL_VERDICT trade plan at the moment this entry was written.
        payload:
          description: Action-specific payload (request/response, fills, errors). Secrets are scrubbed before write.
        prevHash: { type: string }
        hash: { type: string }
      required: [index, ts, missionId, action, source, missionObjective, verdictRef, tradePlan, payload, prevHash, hash]
    LedgerPage:
      type: object
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/LedgerEntry" }
        total: { type: integer }
      required: [entries, total]
    LedgerVerifyResult:
      type: object
      properties:
        ok: { type: boolean }
        totalEntries: { type: integer }
        lastHash: { type: string }
        brokeAtIndex: { type: integer, nullable: true }
        brokeAtReason: { type: string, nullable: true }
        missionEntries: { type: integer, nullable: true }
        missionFirstIndex: { type: integer, nullable: true }
        missionLastIndex: { type: integer, nullable: true }
        missionLastHash: { type: string, nullable: true }
      required: [ok, totalEntries, lastHash, brokeAtIndex, brokeAtReason]
...
    ExecutionConfig:
      type: object
      description: Operator-facing view of the execution config. The private key is never returned — `hasPrivateKey` only indicates whether one is stored.
      properties:
        walletAddress:
          type: string
          nullable: true
          description: Hyperliquid wallet address (0x…40 hex chars). Public — safe to display.
        hasPrivateKey:
          type: boolean
          description: True if an encrypted private key is stored. Real-mode execution is only allowed when this is true.
        useTestnet:
          type: boolean
          description: When true, all signed orders go to api.hyperliquid-testnet.xyz instead of mainnet.
        paperMode:
          type: boolean
          description: When true, no signed orders are sent — fills are simulated against the live mark and persisted to the same execution_orders / execution_positions tables with paper=true.
        notionalPerTradeUsd:
          type: number
          description: USD notional placed at entry. Size = notional / entry_price, rounded to the perp's szDecimals.
        maxNotionalUsd:
          type: number
          description: Hard cap. notionalPerTradeUsd may not exceed this.
        maxConcurrentTrades:
          type: integer
          description: Max number of OPEN positions across all missions. Execution refuses to place a new entry when this cap is hit.
        defaultLeverage:
          type: integer
          description: Cross-margin leverage to send with new orders. (Reserved — current implementation relies on account-level leverage.)
        updatedAt:
          type: string
          format: date-time
      required:
        - hasPrivateKey
        - useTestnet
        - paperMode
        - notionalPerTradeUsd
        - maxNotionalUsd
        - maxConcurrentTrades
        - defaultLeverage
        - updatedAt
    UpdateExecutionConfigBody:
      type: object
      description: "All fields optional. Omit a field to leave it unchanged. Send privateKey as empty string to clear a stored key."
      properties:
        walletAddress:
          type: string
          nullable: true
          pattern: "^0x[a-fA-F0-9]{40}$"
        privateKey:
          type: string
          nullable: true
          description: 32-byte hex (with or without 0x prefix). Encrypted at rest with AES-256-GCM. Empty string clears.
        useTestnet:
          type: boolean
        paperMode:
          type: boolean
        notionalPerTradeUsd:
          type: number
        maxNotionalUsd:
          type: number
        maxConcurrentTrades:
          type: integer
        defaultLeverage:
          type: integer
    ExecutionOrder:
      type: object
      properties:
        id: { type: integer }
        missionId: { type: integer }
        kind:
          type: string
          enum: [entry, stop, tp1, tp2]
        side:
          type: string
          enum: [buy, sell]
        reduceOnly: { type: boolean }
        orderType:
          type: string
          enum: [market, limit, trigger]
        triggerPx: { type: number, nullable: true }
        limitPx: { type: number, nullable: true }
        sz: { type: number }
        hlOrderId:
          type: string
          nullable: true
          description: Hyperliquid order id (oid) returned by the exchange. Null for paper rows and for legs that errored before placement.
        status:
          type: string
          enum: [pending, open, filled, cancelled, rejected, error]
        filledSz: { type: number }
        avgFillPx: { type: number, nullable: true }
        errorMessage: { type: string, nullable: true }
        paper: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - missionId
        - kind
        - side
        - reduceOnly
        - orderType
        - sz
        - status
        - filledSz
        - paper
        - createdAt
        - updatedAt
    ExecutionPosition:
      type: object
      properties:
        id: { type: integer }
        missionId: { type: integer }
        symbol: { type: string }
        side:
          type: string
          enum: [long, short]
        sz: { type: number }
        entryPx: { type: number }
        unrealizedPnlUsd: { type: number }
        realizedPnlUsd: { type: number }
        status:
          type: string
          enum: [open, closed, killed]
        paper: { type: boolean }
        openedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time, nullable: true }
        lastSyncedAt: { type: string, format: date-time }
      required:
        - id
        - missionId
        - symbol
        - side
        - sz
        - entryPx
        - unrealizedPnlUsd
        - realizedPnlUsd
        - status
        - paper
        - openedAt
        - lastSyncedAt
    ExecutionPlacedLeg:
      type: object
      properties:
        kind:
          type: string
          enum: [entry, stop, tp1, tp2]
        side:
          type: string
          enum: [buy, sell]
        reduceOnly: { type: boolean }
        orderType:
          type: string
          enum: [market, trigger]
        triggerPx: { type: number, nullable: true }
        limitPx: { type: number, nullable: true }
        sz: { type: number }
        hlOrderId: { type: string, nullable: true }
        status:
          type: string
          enum: [open, filled, rejected, error]
        filledSz: { type: number }
        avgFillPx: { type: number, nullable: true }
        errorMessage: { type: string, nullable: true }
      required:
        - kind
        - side
        - reduceOnly
        - orderType
        - sz
        - status
        - filledSz
    ExecutionPlacementResult:
      type: object
      properties:
        paper:
          type: boolean
          description: True if executed in paper mode (no signed orders sent).
        legs:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionPlacedLeg"
        position:
          nullable: true
          type: object
          properties:
            symbol: { type: string }
            side:
              type: string
              enum: [long, short]
            sz: { type: number }
            entryPx: { type: number }
          required: [symbol, side, sz, entryPx]
      required:
        - paper
        - legs
    ExecutionKillResult:
      type: object
      properties:
        cancelled:
          type: integer
          description: Number of resting orders successfully cancelled (real-mode count from Hyperliquid; paper-mode = the row count).
        closed:
          type: boolean
          description: True if an open position was flat-closed.
        closedPx:
          type: number
          nullable: true
          description: Mark price (paper) or actual fill price (real) at which the position was closed.
        errors:
          type: array
          items: { type: string }
      required:
        - cancelled
        - closed
        - errors
    MissionExecutionSnapshot:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/ExecutionOrder"
        position:
          nullable: true
          $ref: "#/components/schemas/ExecutionPosition"
      required:
        - orders